+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Intermediate work on Appointments UI [#2464](https://github.com/rokwire/illinois-app/issues/2464).
+### Changed
+- Demographics Questionnaire renamed to Research, various UX updates applied [#2465](https://github.com/rokwire/illinois-app/issues/2465).
 
 ## [4.2.5] - 2022-10-19
 ### Added